--- a/lib/galaxy/objectstore/__init__.py
+++ b/lib/galaxy/objectstore/__init__.py
@@ -125,6 +125,14 @@ class ObjectStore(object):
         """
         raise NotImplementedError()
 
+    def set_object_store_id(self, obj):
+        """
+        Set an object store identifier (``object_store_id``) for the supplied object.
+
+        For simple object stores this may no be needed and so a no-op default
+        method is provided by this class.
+        """
+
     def empty(self, obj, base_dir=None, extra_dir=None, extra_dir_at_root=False, alt_name=None, obj_dir=False):
         """
         Test if the object identified by `obj` has content.
@@ -483,7 +491,13 @@ class NestedObjectStore(ObjectStore):
 
     def create(self, obj, **kwargs):
         """Create a backing file in a random backend."""
-        random.choice(list(self.backends.values())).create(obj, **kwargs)
+        self.set_object_store_id(obj)
+        return self._call_method('create', obj, **kwargs)
+
+    def set_object_store_id(self, obj):
+        """Set the object store ID (backend) to use."""
+        if obj.object_store_id is None:
+            self._random_backend().set_object_store_id(obj)
 
     def empty(self, obj, **kwargs):
         """For the first backend that has this `obj`, determine if it is empty."""
@@ -515,6 +529,9 @@ class NestedObjectStore(ObjectStore):
     def get_object_url(self, obj, **kwargs):
         """For the first backend that has this `obj`, get its URL."""
         return self._call_method('get_object_url', obj, None, False, **kwargs)
+
+    def _random_backend(self):
+        return random.choice(list(self.backends.values()))
 
     def _call_method(self, method, obj, default, default_is_exception,
             **kwargs):
@@ -631,21 +648,26 @@ class DistributedObjectStore(NestedObjectStore):
 
     def create(self, obj, **kwargs):
         """The only method in which obj.object_store_id may be None."""
-        if obj.object_store_id is None or not self.exists(obj, **kwargs):
-            if obj.object_store_id is None or obj.object_store_id not in self.weighted_backend_ids:
-                try:
-                    obj.object_store_id = random.choice(self.weighted_backend_ids)
-                except IndexError:
-                    raise ObjectInvalid('objectstore.create, could not generate '
-                                        'obj.object_store_id: %s, kwargs: %s'
-                                        % (str(obj), str(kwargs)))
+        object_store_id = obj.object_store_id
+        if object_store_id is None or not self.exists(obj, **kwargs):
+            if object_store_id is None or object_store_id not in self.weighted_backend_ids:
+                self.set_object_store_id(obj, **kwargs)
                 _create_object_in_session(obj)
                 log.debug("Selected backend '%s' for creation of %s %s"
                           % (obj.object_store_id, obj.__class__.__name__, obj.id))
             else:
                 log.debug("Using preferred backend '%s' for creation of %s %s"
-                          % (obj.object_store_id, obj.__class__.__name__, obj.id))
+                          % (object_store_id, obj.__class__.__name__, obj.id))
             self.backends[obj.object_store_id].create(obj, **kwargs)
+
+    def set_object_store_id(self, obj, **kwargs):
+        if obj.object_store_id is None:
+            try:
+                obj.object_store_id = random.choice(self.weighted_backend_ids)
+            except IndexError:
+                raise ObjectInvalid('objectstore.create, could not generate '
+                                    'obj.object_store_id: %s, kwargs: %s'
+                                    % (str(obj), str(kwargs)))
 
     def _call_method(self, method, obj, default, default_is_exception, **kwargs):
         object_store_id = self.__get_store_id_for(obj, **kwargs)
@@ -702,7 +724,13 @@ class HierarchicalObjectStore(NestedObjectStore):
 
     def create(self, obj, **kwargs):
         """Call the primary object store."""
-        self.backends[0].create(obj, **kwargs)
+        self._primary_backend().create(obj, **kwargs)
+
+    def set_object_store_id(self, obj, **kwargs):
+        self._primary_backend().set_object_store_id(obj, **kwargs)
+
+    def _primary_backend(self):
+        return self.backends[0]
 
 
 def build_object_store_from_config(config, fsmon=False, config_xml=None):