@jmchilton 2018-06-22
Hey past-@jmchilton - you refactored to eliminate a flush but didn't eliminate the flush so your messy data makes a lot of sense 😠.