@jmchilton 2018-07-11
Updated main comment with fresh times with latest fixed version of this and latest dev. It pretty clearly works as an optimization - for tools with large number of output datasets it gets close to cutting the time to execute the tool setup in half.