@jmchilton 2018-08-01
Closing in favor of the better https://github.com/galaxyproject/galaxy/pull/6550.